fix(layout): 각주 개체 판정을 수집 대상 전체로 넓히고 기본 각주 모양도 확정 상태로 나른다
PR 리뷰 두 건을 합성 문서로 전부 재현한 뒤 고쳤다. 실물 코퍼스(헌법주석 1,030쪽)
렌더는 각주 블록 3,564개가 **좌표까지 완전히 동일**해, 둘 다 잠재 결함만 닫는다.

**각주가 그리는 모든 개체를 판정에 넣는다.** "이 각주가 개체를 담는가"(쪽 끝 분할
금지·CT 높이 보존)를 예약은 높이 하한 술어 `hasFloatingObject`로 판정했는데, 그 술어는
글 앞으로·글 뒤로 배치한 오버레이 개체를 뺀다 — 흐름을 안 점유하고 줄에도 안 담기기
때문이다. 배치는 그 개체를 수집해 CT 높이를 지키므로 위치 리셋 캐시 각주에서 예약
58.36pt·배치 78.2pt로 갈렸고, 여러 문단 각주에서는 앞 문단이 캐시 합(64 → 46.88pt)으로
줄어 그림과 겹쳤다. 수집기 관문(표·`handledControl`·`collectible`)을 그대로 되풀이한
`hasCollectibleObject`를 두고 예약과 배치가 같은 술어를 쓴다. `hasFloatingObject`는 표
셀의 하한 재수집 게이트로만 남는다.

**기본 각주 모양으로 잰 것도 확정 상태다.** 공개 `place`의 첫 호출에서 각주 모양을
생략하면 기본 모양으로 조판한 뒤에도 이월 입력에 nil이 남아, 다음 호출이 다른 모양을
지정하면 그 모양을 새로 채택해 `placedLength`의 기준 문자열이 달라졌다 — `1`에서
`(1)`로 바뀐 재현에서 앞 조각의 마지막 두 글자가 겹쳐 그려졌다. `Input.measuredShape`를
`MeasuredShape?`로 두어 "아직 안 잼"(nil)과 "기본 모양으로 잼"(`MeasuredShape(nil)`)을
가른다.

파일 길이 상한(SwiftLint error)에 닿아 개체 배치 술어를 `HwpParagraphObjectPredicates`
로, 리뷰 재현 테스트를 `HwpFootnoteContinuationReviewTests`로, 합성 조립 헬퍼를
`FootnoteContinuationSupport`로 갈랐다.

검증: 새 재현 3종(오버레이 개체의 예약≠배치·뒤 문단 오버레이의 앞 문단 높이·기본
모양 이월 중복)이 수정 전 실패·수정 후 통과. 전체 `swift test` 4번들 통과(CoreHwp
1,622·HwpKitCore 848·Native 281·HwpKit 192). 코퍼스 재렌더는 쪽수 1,030·겹침 0쪽·
한글 PDF 대조 1,027/1,030으로 수정 전과 같고, 각주 블록 diff가 0이다.
