+++ Sources/HwpKitCore/AGENTS.md
@@ -1295,7 +1295,11 @@ paraShape와 같은 값**이어야 한다.
   같아도 `placedLength`가 다른 글자를 가리킨다 (실측: `(1)` → `1`에서 `word45`가 `rd45`로
   잘림). 예약도 그 모양으로 쟀으므로 같은 값을 써야 예약 ≡ 배치가 성립한다 —
   `sizeResolver`(R44 #1)와 같은 규약이고, `measure`가 처음 잰 값을 각인해 나른다.
-  구분선 여백·길이는 **쪽**의 몫이라 그대로 현재 구역 모양을 쓴다. 경계가 줄 중간에 떨어지면 그 줄의 앞부분을
+  각인은 **"아직 안 잼"과 "기본 모양(nil)으로 잼"을 가른다** (`Input.measuredShape` —
+  `MeasuredShape?`의 이중 optional): 둘을 같은 nil로 두면 첫 `place`에서 모양을 생략해
+  기본 모양으로 잰 이월 입력이 다음 쪽의 모양을 새로 채택해 문자열이 밀린다 (실측: `1` →
+  `(1)`에서 앞 조각의 마지막 두 글자 중복). 구분선 여백·길이는 **쪽**의 몫이라 그대로
+  현재 구역 모양을 쓴다. 경계가 줄 중간에 떨어지면 그 줄의 앞부분을
   잘라 내되 줄 기하와 앵커는 앞 쪽 폭 기준이라 근사다 (분할 각주는 개체를 안 담으므로
   앵커는 버린다).
   **개체를 담은 각주는 나누지 않는다** (`NoteMeasurement.carriesObjects`): 개체는 앵커
@@ -1307,9 +1311,13 @@ paraShape와 같은 값**이어야 한다.
   겹친다 (실측: 블록 44.16pt에 개체 하단 79.35pt). **두 판정의 범위는 각주 단위로 같다**
   (#165 리뷰): 문단 단위로 보면 개체가 뒤 문단에만 있을 때 앞 문단만 캐시 합으로 줄어 그
   문단의 마지막 글줄 위로 뒤 문단의 그림이 올라온다. 판정은 `measure`가 `noteId`별로 한 번
-  (`hasFloatingObject` 상위집합 술어) 하고, **예약도 같은 범위를 본다** — 개체를 담은
+  (`hasCollectibleObject` — 수집기 관문을 되풀이한 술어) 하고, **예약도 같은 범위를 본다** — 개체를 담은
   각주의 문단은 그 문단에 개체가 없어도 캐시 빠른 길(`measuredFootnoteTextHeight`)로 가지
-  않는다 (가면 예약이 캐시 합, 배치가 CT 높이로 갈린다). 여러 문단짜리 각주의 뒤 문단이 0에서 다시 시작하면 그
+  않는다 (가면 예약이 캐시 합, 배치가 CT 높이로 갈린다). **술어는 수집 대상 전체다**
+  (#165 리뷰): 높이 하한 술어 `hasFloatingObject`는 글 앞으로·글 뒤로 개체(오버레이 —
+  흐름을 안 점유하고 줄에도 안 담긴다)를 빼는데 배치는 그 개체를 수집해 CT 높이를
+  지키므로, 예약만 하한 술어로 판정하면 같은 각주를 다르게 본다 (실측: 예약 58.36 vs 배치
+  78.2pt). `hasFloatingObject`는 표 셀의 하한 재수집 게이트로만 남는다. 여러 문단짜리 각주의 뒤 문단이 0에서 다시 시작하면 그
   문단부터 통째로 다음 쪽이다. 들어맞음 허용 오차는 부동소수 잡음뿐 — 한글은 1 HWPUNIT
   여유에도 싣는다. 진행 보장: 빈 쪽에도 안 들어가는 분할 지점 없는 각주는 그대로 싣는다
   (코퍼스 0건). 결과: 헌법주석 겹침 366쪽 → **0쪽**, 쪽수 1,030 그대로, 이어짐 쪽 64
