+++ Tests/HwpKitCoreTests/HwpFootnoteContinuationTests.swift
@@ -6,163 +6,8 @@ import Nimble
 import XCTest
 
 #if canImport(CoreText)
-    /// 합성 조립·페이지네이션 헬퍼 — 클래스 본문이 SwiftLint type_body_length
-    /// 상한(400줄)에 닿아 갈라 뒀다.
-    private enum FootnoteContinuationSupport {
-        static let pitch: CGFloat = 11.72
-        static let box: CGFloat = 9.0
-        static let overhead: CGFloat = HwpRenderTuning.Footnote.dividerDefaultMarginTop
-            + HwpRenderTuning.Footnote.dividerDefaultMarginBottom
-
-        // MARK: - 합성 조립
-
-        static func lineSegPayload(
-            _ lines: [(location: Int32, height: Int32, spacing: Int32)],
-            property: UInt32 = 0x60000
-        ) -> Data {
-            var payload = Data()
-            for line in lines {
-                withUnsafeBytes(of: UInt32(0).littleEndian) { payload.append(contentsOf: $0) }
-                withUnsafeBytes(of: line.location.littleEndian) { payload.append(contentsOf: $0) }
-                withUnsafeBytes(of: line.height.littleEndian) { payload.append(contentsOf: $0) }
-                withUnsafeBytes(of: line.height.littleEndian) { payload.append(contentsOf: $0) }
-                withUnsafeBytes(of: Int32(765).littleEndian) { payload.append(contentsOf: $0) }
-                withUnsafeBytes(of: line.spacing.littleEndian) { payload.append(contentsOf: $0) }
-                withUnsafeBytes(of: Int32(0).littleEndian) { payload.append(contentsOf: $0) }
-                withUnsafeBytes(of: Int32(42520).littleEndian) { payload.append(contentsOf: $0) }
-                withUnsafeBytes(of: property.littleEndian) { payload.append(contentsOf: $0) }
-            }
-            return payload
-        }
-
-        /// 각주 줄 캐시 — 세로 위치 목록 (HWPUNIT, 각주 시작 기준). 줄 높이 900·간격 272.
-        static func noteLines(
-            _ locations: [Int32]
-        ) -> [(location: Int32, height: Int32, spacing: Int32)] {
-            locations.map { (location: $0, height: 900, spacing: 272) }
-        }
-
-        /// 자동 번호 마커로 시작하는 각주 첫 문단 + 줄 캐시. 줄바꿈 문자로 CT 줄 수를
-        /// 캐시 줄 수와 같게 고정한다 (조각 문자열 대응이 비례 근사라 어긋나면 흔들린다).
-        static func note(
-            lines texts: [String], locations: [Int32]
-        ) throws -> CoreHwp.HwpParagraph {
-            var paragraph = HwpSynthetic.noteParagraph(
-                " " + texts.joined(separator: "\n"),
-                autoNumber: HwpSynthetic.autoNumberControl(kind: 1, decorationTail: ")")
-            )
-            paragraph.paraLineSeg = try CoreHwp.HwpParaLineSeg.load(
-                lineSegPayload(noteLines(locations))
-            )
-            return paragraph
-        }
-
-        /// 각주의 뒤 문단 (마커 없음) + 줄 캐시.
-        static func notePlainParagraph(
-            _ text: String, locations: [Int32]
-        ) throws -> CoreHwp.HwpParagraph {
-            var paragraph = try HwpSynthetic.textParagraph(text)
-            paragraph.paraLineSeg = try CoreHwp.HwpParaLineSeg.load(
-                lineSegPayload(noteLines(locations))
-            )
-            return paragraph
-        }
-
-        /// 각주 참조 마커를 `noteCount`개 가진 한 줄 본문 문단 — 세로 위치 `location`.
-        static func host(
-            at location: Int32, notes: [[CoreHwp.HwpParagraph]]
-        ) throws -> CoreHwp.HwpParagraph {
-            var host = try HwpSynthetic.splitParagraphWithMixedMarkers(
-                lines: [(characters: 5, markers: Array(repeating: 17, count: notes.count))],
-                segments: [(location: location, height: 1000, textStart: 0)]
-            )
-            host.ctrlHeaderArray = notes.map {
-                .footnote(HwpSynthetic.listControl(ctrlId: .footnote, paragraphs: $0))
-            }
-            return host
-        }
-
-        /// 절대 캐시 모드 문서 — 첫 loc > 0인 캐시 문단이 다수여야 한다 (감지 규칙).
-        static func paginate(
-            _ body: [CoreHwp.HwpParagraph],
-            footnoteNumberingMode: UInt32 = 0
-        ) -> HwpPaginator {
-            let section = HwpSynthetic.section(
-                firstParagraphControls: [.section(HwpSynthetic.sectionDef(
-                    footnoteNumberingMode: footnoteNumberingMode
-                ))],
-                bodyParagraphs: body
-            )
-            return HwpPaginator(
-                sections: [section],
-                index: HwpIndex(from: CoreHwp.HwpFile()),
-                fontResolver: .testDeterministic
-            )
-        }
-
-        /// 다음 쪽 본문 — 세로 위치가 되돌아가 한글의 쪽 절단점이 된다.
-        static func nextPageBody(bottomAt location: Int32? = nil) throws -> [CoreHwp.HwpParagraph] {
-            var body = [try HwpSynthetic.lineSegParagraph(
-                "다음 쪽 문단", segments: [(location: 2720, height: 1500)]
-            )]
-            if let location {
-                body.append(try HwpSynthetic.lineSegParagraph(
-                    "다음 쪽 아래 문단", segments: [(location: location, height: 1000)]
-                ))
-            }
-            return body
-        }
-
-        static func footnoteBlocks(on page: HwpPage?) -> [HwpFootnoteBlock] {
-            (page?.blocks ?? []).compactMap { block in
-                if case let .footnote(note) = block.payload {
-                    return note
-                }
-                return nil
-            }
-        }
-
-        static func text(_ block: HwpFootnoteBlock) -> String {
-            block.paragraphs.map(\.attributedString.string).joined()
-        }
-
-        /// 앞·뒤 장식 문자만 지정한 각주 모양 — 구역마다 번호 라벨이 달라지는 문서를
-        /// 재현한다. 구분선 값은 건드리지 않아 기본 여백이 그대로 쓰인다.
-        static func footnoteShape(
-            head: Character?, tail: Character?
-        ) -> CoreHwp.HwpFootnoteShape {
-            var shape = CoreHwp.HwpSectionDef().footNoteShape
-            shape.decorationHeadRawValue = head?.utf16.first ?? 0
-            shape.decorationTailRawValue = tail?.utf16.first ?? 0
-            return shape
-        }
-
-        /// 지정 폭의 A4 쪽 기하 (단위 테스트용) — 구역이 바뀌며 폭이 달라지는 문서를
-        /// 페이지네이터 없이 재현한다.
-        static func geometry(contentWidth: CGFloat) -> HwpPageGeometry {
-            let frame = CGRect(x: 72, y: 72, width: contentWidth, height: 698)
-            return HwpPageGeometry(
-                pageSize: CGSize(width: 595, height: 842),
-                margins: HwpPageMargins(top: 72, left: 72, bottom: 72, right: 72),
-                contentFrame: frame,
-                headerFrame: nil,
-                footerFrame: nil,
-                columnFrames: [frame]
-            )
-        }
-
-        static func contentBottom(of page: HwpPage) -> CGFloat {
-            page.size.height - page.margins.bottom
-        }
-
-        /// host 줄 상자 아래에서 본문 하단까지 `available`pt (구분선 여백 제외) 를 남기는
-        /// host 세로 위치. 본문 상단 56.68pt·하단 799.36pt (sectionDef 기본 쪽).
-        static func hostLocation(leaving available: CGFloat) -> Int32 {
-            let contentHeight: CGFloat = 841.88 - 56.68 - 42.52
-            let boxBottom = contentHeight - overhead - available
-            return Int32((boxBottom * 100).rounded()) - 1000
-        }
-    }
+    // 합성 조립·페이지네이션 헬퍼 — 클래스 본문이 SwiftLint type_body_length
+    // 상한(400줄)에 닿아 갈라 뒀다.
 
     /// 절대 캐시 모드의 각주 이어짐 (#165) — 한글.app 12.30 실측 규칙
     /// (`HwpFootnoteContinuation.swift` 머리 주석) 을 합성 문서로 잠근다.
@@ -369,227 +214,6 @@ import XCTest
             // 개체는 옮겨진 각주 하나에만 있고 그 블록 안에 있다.
             expect(moved.shapes.count) == 1
             expect(moved.shapes.first?.rect.maxY) <= moved.frame.height + 0.01
-        }
-
-        /// 폭이 다른 구역으로 이월돼도 **앞 쪽이 소비한 문자 경계**가 보존돼야 한다
-        /// (#165 리뷰). 이어지는 조각은 새 쪽의 폭으로 문단을 다시 조판하므로 줄 나눔이
-        /// 달라진다 — 캐시 줄 인덱스를 CT 줄 수에 비례 환산한 경계는 앞 쪽이 실제로 그린
-        /// 마지막 글자와 다른 자리를 가리켜 글자가 통째로 사라지거나 겹친다.
-        func testContinuationKeepsConsumedTextAcrossAWidthChange() throws {
-            let words = (1 ... 60).map { "word\($0)" }.joined(separator: " ")
-            var paragraph = HwpSynthetic.noteParagraph(
-                " " + words,
-                autoNumber: HwpSynthetic.autoNumberControl(kind: 1, decorationTail: ")")
-            )
-            paragraph.paraLineSeg = try CoreHwp.HwpParaLineSeg.load(
-                Support.lineSegPayload(Support.noteLines([0, 1172, 2344, 0, 1172, 2344]))
-            )
-            let input = HwpFootnoteLayout.Input(paragraph: paragraph, number: 1)
-            let layout = HwpFootnoteLayout(fontResolver: .testDeterministic)
-            let index = HwpIndex(from: CoreHwp.HwpFile())
-            let wide = Support.geometry(contentWidth: 451)
-            let narrow = Support.geometry(contentWidth: 200)
-
-            // 통째 배치(빈 쪽)의 문자열이 이 각주의 전체 텍스트다 — 폭과 무관하다.
-            let whole = layout.place(
-                footnotes: [input], onPage: wide, index: index, limitsAreaToHalfContent: false
-            )
-            let full = Support.text(try XCTUnwrap(whole.blocks.first))
-            // 앞 세 줄(44.16pt)은 들어가고 전체(91.04pt)는 안 들어가는 자리에서 나눈다.
-            let split = layout.place(
-                footnotes: [input], onPage: wide, index: index,
-                limitsAreaToHalfContent: false, bodyBottom: wide.contentFrame.maxY - 64.2
-            )
-            let head = Support.text(try XCTUnwrap(split.blocks.first))
-            expect(split.overflow.count) == 1
-            // 이월분은 **좁은** 쪽에서 다시 조판된다.
-            let carried = layout.place(
-                footnotes: split.overflow, onPage: narrow, index: index,
-                limitsAreaToHalfContent: false
-            )
-            let tail = Support.text(try XCTUnwrap(carried.blocks.first))
-
-            expect(full.hasPrefix(head)) == true
-            expect(full.hasSuffix(tail)) == true
-            expect(head.count + tail.count) == full.count
-        }
-
-        /// 개체를 담은 각주를 통째로 배치할 때 블록이 그 개체를 담아야 한다 (#165 리뷰).
-        /// 쪽에 걸친 캐시(비단조)의 쪽 몫 합은 한글이 **두 쪽에 나눠** 그린 높이인데,
-        /// 개체를 담은 각주는 나누지 않고 한 쪽에 통째로 그리므로 그 합으로는 CT 좌표로
-        /// 수집한 개체를 담지 못해 다음 각주와 겹친다.
-        func testWholeNoteWithAnObjectKeepsAHeightThatContainsIt() throws {
-            var paragraph = CoreHwp.HwpParagraph()
-            var paraText = CoreHwp.HwpParaText()
-            paraText.charArray = [CoreHwp.HwpChar(type: .extended, value: 18)]
-                + " 첫째 줄\n둘째 줄\n셋째 줄\n넷째 줄 ".utf16
-                .map { CoreHwp.HwpChar(type: .char, value: $0) }
-                + [CoreHwp.HwpChar(type: .extended, value: 11)]
-            paragraph.paraText = paraText
-            paragraph.ctrlHeaderArray = [
-                HwpSynthetic.autoNumberControl(kind: 1, decorationTail: ")"),
-                .genShapeObject(HwpSynthetic.inlineShapeObject(width: 2000, height: 3000)),
-            ]
-            // 세로 위치가 되돌아가는 캐시 — 한글이 두 쪽에 나눠 그린 각주다.
-            paragraph.paraLineSeg = try CoreHwp.HwpParaLineSeg.load(
-                Support.lineSegPayload(Support.noteLines([0, 1172, 0, 1172]))
-            )
-            let layout = HwpFootnoteLayout(fontResolver: .testDeterministic)
-            let placement = layout.place(
-                footnotes: [HwpFootnoteLayout.Input(paragraph: paragraph, number: 1)],
-                onPage: Support.geometry(contentWidth: 451),
-                index: HwpIndex(from: CoreHwp.HwpFile()),
-                limitsAreaToHalfContent: false
-            )
-            let block = try XCTUnwrap(placement.blocks.first)
-            let objectBottom = block.shapes.map(\.rect.maxY).max() ?? 0
-            expect(objectBottom) > 0
-            expect(block.frame.height) >= objectBottom - 0.01
-        }
-
-        /// 쪽마다 번호를 새로 시작하면 (표 134 모드 2) 이월된 각주와 새 각주가 **같은 표시
-        /// 번호**를 갖는다 (#165 리뷰). 번호로 그룹을 나누면 둘이 한 각주로 합쳐져 사이
-        /// 여백이 사라지고, 앞 각주의 마지막 줄 간격까지 높이에 남는다.
-        func testCarriedAndNewNoteSharingADisplayNumberStayDistinct() async throws {
-            let split = try Support.note(lines: ["첫째 줄", "둘째 줄"], locations: [0, 0])
-            let fresh = try Support.note(lines: ["새 각주"], locations: [0])
-            let host = try Support.host(at: Support.hostLocation(leaving: 15), notes: [[split]])
-            let nextHost = try Support.host(at: 2720, notes: [[fresh]])
-            let paginator = Support.paginate([host, nextHost], footnoteNumberingMode: 2)
-
-            let second = try await paginator.page(at: 1)
-            let blocks = Support.footnoteBlocks(on: second)
-            expect(blocks.count) == 2
-            expect(blocks.map(\.number)) == [1, 1]
-            // 이어지는 조각은 그 각주의 끝이다 — 마지막 줄의 줄 간격을 세지 않는다.
-            expect(blocks[0].frame.height).to(beCloseTo(Support.box, within: 0.01))
-            // 서로 다른 각주 사이에는 여백이 들어간다.
-            expect(blocks[1].frame.minY - blocks[0].frame.maxY).to(beCloseTo(
-                HwpRenderTuning.Footnote.dividerDefaultSpacingBetweenNotes, within: 0.01
-            ))
-        }
-
-        /// 구역이 바뀌어 **번호 장식**이 달라져도 소비한 문자 경계가 보존돼야 한다
-        /// (#165 리뷰). 이어지는 조각은 그 쪽 구역의 각주 모양으로 문자열을 다시 만드는데,
-        /// 라벨 길이가 달라지면 본문 글자 위치가 통째로 밀려 앞 쪽이 그린 마지막 글자와
-        /// 다른 자리에서 이어진다 — 폭이 같아도 글자가 잘린다.
-        func testContinuationKeepsConsumedTextAcrossAShapeChange() throws {
-            let words = (1 ... 60).map { "word\($0)" }.joined(separator: " ")
-            var paragraph = HwpSynthetic.noteParagraph(
-                " " + words,
-                // 장식을 안 실은 자동 번호 — 라벨이 **구역 각주 모양**을 따른다.
-                autoNumber: HwpSynthetic.autoNumberControl(kind: 1)
-            )
-            paragraph.paraLineSeg = try CoreHwp.HwpParaLineSeg.load(
-                Support.lineSegPayload(Support.noteLines([0, 1172, 2344, 0, 1172, 2344]))
-            )
-            let input = HwpFootnoteLayout.Input(paragraph: paragraph, number: 1)
-            let layout = HwpFootnoteLayout(fontResolver: .testDeterministic)
-            let index = HwpIndex(from: CoreHwp.HwpFile())
-            let page = Support.geometry(contentWidth: 451)
-            let decorated = Support.footnoteShape(head: "(", tail: ")")
-            let plain = Support.footnoteShape(head: nil, tail: nil)
-
-            let whole = layout.place(
-                footnotes: [input], onPage: page, index: index,
-                footnoteShape: decorated, limitsAreaToHalfContent: false
-            )
-            let full = Support.text(try XCTUnwrap(whole.blocks.first))
-            // 앞 세 줄(32.44pt)은 들어가고 전체(67.6pt)는 안 들어가는 자리에서 나눈다.
-            let split = layout.place(
-                footnotes: [input], onPage: page, index: index,
-                footnoteShape: decorated, limitsAreaToHalfContent: false,
-                bodyBottom: page.contentFrame.maxY - 54.2
-            )
-            let head = Support.text(try XCTUnwrap(split.blocks.first))
-            expect(split.overflow.count) == 1
-            // 이월분은 **장식 없는** 구역에서 다시 조판된다.
-            let carried = layout.place(
-                footnotes: split.overflow, onPage: page, index: index,
-                footnoteShape: plain, limitsAreaToHalfContent: false
-            )
-            let tail = Support.text(try XCTUnwrap(carried.blocks.first))
-
-            expect(full.hasPrefix(head)) == true
-            expect(full.hasSuffix(tail)) == true
-            expect(head.count + tail.count) == full.count
-        }
-
-        /// 개체가 **뒤 문단**에 붙은 각주도 통째로 CT 높이를 쓴다 (#165 리뷰). 분할 금지는
-        /// 각주 단위인데 높이 보존이 문단 단위면, 앞 문단이 캐시 합으로 줄어 그 문단의 마지막
-        /// 글줄 위로 뒤 문단의 그림이 올라온다.
-        func testObjectInALaterParagraphKeepsEveryParagraphOnCTHeight() throws {
-            let text = try Support.note(
-                lines: ["첫째 줄", "둘째 줄", "셋째 줄", "넷째 줄"], locations: [0, 1172, 0, 1172]
-            )
-            var picture = CoreHwp.HwpParagraph()
-            var paraText = CoreHwp.HwpParaText()
-            paraText.charArray = "그림 ".utf16.map { CoreHwp.HwpChar(type: .char, value: $0) }
-                + [CoreHwp.HwpChar(type: .extended, value: 11)]
-            picture.paraText = paraText
-            picture.ctrlHeaderArray = [
-                .genShapeObject(HwpSynthetic.inlineShapeObject(width: 2000, height: 3000)),
-            ]
-            picture.paraLineSeg = try CoreHwp.HwpParaLineSeg.load(
-                Support.lineSegPayload(Support.noteLines([0]))
-            )
-            let layout = HwpFootnoteLayout(fontResolver: .testDeterministic)
-            let placement = layout.place(
-                footnotes: [
-                    HwpFootnoteLayout.Input(paragraph: text, number: 1),
-                    HwpFootnoteLayout.Input(paragraph: picture, number: 1),
-                ],
-                onPage: Support.geometry(contentWidth: 451),
-                index: HwpIndex(from: CoreHwp.HwpFile()),
-                limitsAreaToHalfContent: false
-            )
-            expect(placement.blocks.count) == 2
-            let first = try XCTUnwrap(placement.blocks.first)
-            // 네 줄 × 16pt — 캐시 합 46.88pt로 줄면 넷째 줄이 그림 아래로 밀린다.
-            expect(first.frame.height).to(beCloseTo(64, within: 1))
-            expect(placement.blocks[1].frame.minY).to(beCloseTo(first.frame.maxY, within: 0.01))
-        }
-
-        /// 표시 번호가 같은 서로 다른 각주는 **식별자 없이는** 한 각주로 묶인다 (#165 리뷰) —
-        /// 남은 자리에 들어가는 이월 조각까지 새 각주와 함께 다음 쪽으로 밀린다.
-        func testNotesSharingADisplayNumberMergeWithoutAnIdentity() throws {
-            let layout = HwpFootnoteLayout(fontResolver: .testDeterministic)
-            let index = HwpIndex(from: CoreHwp.HwpFile())
-            let page = Support.geometry(contentWidth: 451)
-            let split = try Support.note(lines: ["첫째 줄", "둘째 줄"], locations: [0, 0])
-            // 캐시 없는 새 각주 — 나눌 지점이 없어 그룹이 통째로 밀린다.
-            let fresh = HwpSynthetic.noteParagraph(
-                " 새 각주",
-                autoNumber: HwpSynthetic.autoNumberControl(kind: 1, decorationTail: ")")
-            )
-            let first = layout.place(
-                footnotes: [HwpFootnoteLayout.Input(paragraph: split, number: 1)],
-                onPage: page, index: index, limitsAreaToHalfContent: false,
-                bodyBottom: page.contentFrame.maxY - 29.2
-            )
-            let carried = try XCTUnwrap(first.overflow.first)
-
-            // 이월 조각 9.0pt는 남은 12pt에 들어가지만, 같은 번호라 새 각주와 한 그룹이 된다.
-            let merged = layout.place(
-                footnotes: [carried, HwpFootnoteLayout.Input(paragraph: fresh, number: 1)],
-                onPage: page, index: index, limitsAreaToHalfContent: false,
-                bodyBottom: page.contentFrame.maxY - 26.2
-            )
-            expect(merged.blocks).to(beEmpty())
-            expect(merged.overflow.count) == 2
-
-            // 식별자를 주면 서로 다른 각주로 갈려 이월 조각이 남은 자리에 실린다.
-            let distinct = layout.place(
-                footnotes: [
-                    carried,
-                    HwpFootnoteLayout.Input(paragraph: fresh, number: 1, noteId: 2),
-                ],
-                onPage: page, index: index, limitsAreaToHalfContent: false,
-                bodyBottom: page.contentFrame.maxY - 26.2
-            )
-            expect(distinct.blocks.count) == 1
-            expect(Support.text(try XCTUnwrap(distinct.blocks.first))).to(contain("둘째 줄"))
-            expect(distinct.overflow.count) == 1
         }
 
         /// 진행 보장 — 빈 쪽에도 안 들어가는 각주 (분할 지점 없음) 는 참조 쪽에 그대로
